+++ app/+not-found.tsx
@@ -1,4 +1,5 @@
 import { Link, Stack } from 'expo-router';
+import React from 'react';
 import { Image, View } from 'react-native';
 
 import FText from '~/components/Text/FText';
