+++ app/details.tsx
@@ -1,4 +1,5 @@
 import { Stack, useLocalSearchParams, Link } from 'expo-router';
+import React from 'react';
 import { View } from 'react-native';
 
 import FText from '~/components/Text/FText';
